--- integration/src/test/resources/non-blocking-partly-supported-test-config.xml
@@ -1,42 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<mule xmlns="http://www.mulesoft.org/schema/mule/core"
-      xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-      xmlns:test="http://www.mulesoft.org/schema/mule/test"
-      xsi:schemaLocation="
-               http://www.mulesoft.org/schema/mule/core http://www.mulesoft.org/schema/mule/core/current/mule.xsd
-               http://www.mulesoft.org/schema/mule/test http://www.mulesoft.org/schema/mule/test/current/mule-test.xsd">
-
-    <configuration>
-        <expression-language>
-            <global-functions>
-                def assertPayload(){payload == org.mule.tck.junit4.AbstractMuleContextTestCase.TEST_MESSAGE}
-            </global-functions>
-        </expression-language>
-    </configuration>
-
-    <flow name="wiretap">
-        <test:assert-intercepting message="before" responseSameTask="false" responseExpression="#[mel:assertPayload()]"/>
-        <wire-tap>
-            <processor-chain>
-                <set-variable variableName="beforeThread" value="#[mel:Thread.currentThread()]"/>
-                <test:non-blocking-processor/>
-                <test:assert count="1" expression="#[mel:Thread.currentThread()==beforeThread  and assertPayload()]"/>
-            </processor-chain>
-        </wire-tap>
-        <test:non-blocking-processor/>
-    </flow>
-
-    <flow name="childDefaultFlow">
-        <test:assert-intercepting responseSameTask="false" responseExpression="#[mel:assertPayload()]" message="1"/>
-        <expression-component>mel:flowVars['requestThread'] = Thread.currentThread()</expression-component>
-        <flow-ref name="childDefaultFlowChild"/>
-        <test:assert expression="#[mel:assertPayload()]"/>
-        <test:non-blocking-processor/>
-    </flow>
-    <flow name="childDefaultFlowChild" processingStrategy="synchronous">
-        <test:assert-intercepting responseSameTask="true" responseExpression="#[mel:assertPayload()]" message="2"/>
-        <test:non-blocking-processor/>
-        <test:assert expression="#[mel:assertPayload()]"/>
-    </flow>
-
-</mule>
